reorganized files and refactor reply widget
