+++ lib/help_desk/help_desk_main/view/user/page/help_desk_main_desktop_widget.dart
@@ -5,7 +5,7 @@ import 'package:senior_project/assets/font_style.dart';
 import 'package:senior_project/core/template_desktop/view/page/template_desktop.dart';
 import 'package:senior_project/core/template_mobile/view/template_menu_mobile.dart';
 import 'package:senior_project/core/view_model/app_view_model.dart';
-import 'package:senior_project/help_desk/help_desk_main/core/widget/mobile/mobile_widget.dart';
+import 'package:senior_project/help_desk/help_desk_main/view/widget/mobile/mobile_widget.dart';
 import 'package:senior_project/help_desk/help_desk_main/view/user/widget/help_desk_desktop_header_widget.dart';
 import '../widget/help_desk_desktop_body_widget.dart';
