+++ lib/help_desk/help_desk_reply/view/desktop/widget/body_reply_desktop.dart
@@ -2,12 +2,36 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:senior_project/assets/color_constant.dart';
 import 'package:senior_project/core/view_model/app_view_model.dart';
-import 'package:senior_project/help_desk/help_desk_main/model/help_desk_reply_demo.dart';
 import 'package:senior_project/help_desk/help_desk_reply/view/desktop/widget/chat_input_desktop.dart';
 import 'package:senior_project/help_desk/help_desk_reply/view/desktop/widget/description_desktop.dart';
+import 'package:senior_project/help_desk/help_desk_reply/view/widget/message.dart';
 
-class BodyReplyDesktop extends StatelessWidget {
+class BodyReplyDesktop extends StatefulWidget {
   const BodyReplyDesktop({super.key});
+
+  @override
+  State<BodyReplyDesktop> createState() => _BodyReplyDesktopState();
+}
+
+class _BodyReplyDesktopState extends State<BodyReplyDesktop> {
+  List<Map<String, dynamic>> data = [
+    {
+      "text": "Hi people",
+      "isSender": true
+    },
+    {
+      "text": "Okay, How r u?",
+      "isSender": false
+    },
+    {
+      "text": "Fine",
+      "isSender": true
+    },
+    {
+      "text": "asddsssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss",
+      "isSender": false
+    }
+  ];
 
   @override
   Widget build(BuildContext context) {
@@ -65,9 +89,10 @@ class BodyReplyDesktop extends StatelessWidget {
                     child: Padding(
                       padding: const EdgeInsets.symmetric(horizontal: 16),
                       child: ListView.builder(
-                        itemCount: demoMessage.length,
+                        itemCount: data.length,
                         itemBuilder: (context, index) => Message(
-                          message: demoMessage[index],
+                          isSender:  data[index]["isSender"],
+                          text: data[index]["text"],
                         ),
                       ),
                     ),
@@ -84,86 +109,6 @@ class BodyReplyDesktop extends StatelessWidget {
             )
           )
         ],
-      ),
-    );
-  }
-}
-
-class Message extends StatelessWidget {
-  const Message({
-    Key? key,
-    required this.message,
-  }) : super(key: key);
-
-  final MessageChat message;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 24),
-      child: Row(
-        mainAxisAlignment:
-            message.isSender ? MainAxisAlignment.end : MainAxisAlignment.start,
-        children: [
-          if (!message.isSender) ...[
-            //TODO Pull image profile sender
-            const Icon(
-              Icons.face_rounded,
-              color: ColorConstant.red40,
-            )
-          ],
-          const SizedBox(width: 8),
-          TextMessage(
-            messages: message,
-          ),
-          //status
-          if (message.isSender) const MessageStatusDot()
-        ],
-      ),
-    );
-  }
-}
-
-class TextMessage extends StatelessWidget {
-  const TextMessage({
-    Key? key,
-    required this.messages,
-  }) : super(key: key);
-
-  final MessageChat messages;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        // margin: const EdgeInsets.only(top: 24),
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-        decoration: BoxDecoration(
-            color:
-                ColorConstant.orange30.withOpacity(messages.isSender ? 1 : 0.4),
-            borderRadius: BorderRadius.circular(24)),
-        child: Text(
-          messages.text,
-          style:
-              const TextStyle(color: ColorConstant.whiteBlack80, fontSize: 16),
-        ));
-  }
-}
-
-class MessageStatusDot extends StatelessWidget {
-  const MessageStatusDot({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(left: 8),
-      height: 12,
-      width: 12,
-      decoration: const BoxDecoration(
-          color: ColorConstant.green40, shape: BoxShape.circle),
-      child: const Icon(
-        Icons.done_rounded,
-        size: 8,
-        color: ColorConstant.white,
       ),
     );
   }
